Move LibSqlConnection from spin-sqlite to spin-sqlite-libsql
