+++ crates/sqlite-libsql/Cargo.toml
@@ -11,6 +11,7 @@ async-trait = "0.1.68"
 # libsqlite3-sys as used by spin-sqlite-inproc.
 libsql = { version = "0.3.2", features = ["remote"], default-features = false }
 rusqlite = { version = "0.29.0", features = ["bundled"] }
+spin-factor-sqlite = { path = "../factor-sqlite" }
 spin-world = { path = "../world" }
 sqlparser = "0.34"
 tokio = { version = "1", features = ["full"] }
